Only try to save world when authoritative

+++ core/src/main/kotlin/no/elg/infiniteBootleg/world/world/World.kt
@@ -1132,7 +1132,7 @@ abstract class World(
     } finally {
       chunksLock.writeLock().unlock()
     }
-    if (!transientWorld) {
+    if (Main.isAuthoritative && !transientWorld) {
       val worldFolder = worldFolder
       if (worldFolder != null && worldFolder.isDirectory) {
         if (!deleteLockFile(uuid)) {
